Candy and gum classes created

+++ src/main/java/com/techelevator/Purchasable.java
@@ -3,6 +3,6 @@ package com.techelevator;
 public interface Purchasable {
 	
 	 String getName();
-	 int getPrice();
+	 double getPrice();
 
 }
